refactoring to mdbreact
