+++ client/src/assets/styles/styleSam.scss
@@ -24,3 +24,8 @@
   margin: auto;
   text-align: center;
 }
+
+.testButton {
+  width: 100%;
+  margin: auto;
+}
