+++ client/src/components/FeedPageList.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import Button from "@material-ui/core/Button";
+import TextField from "@material-ui/core/TextField";
+import { makeStyles } from "@material-ui/core/styles";
+import Container from "@material-ui/core/Container";
+
+const FeedPageList = props => <div>hi</div>;
+
+export default FeedPageList;
